requirements.txt
mysqlclient==2.2.4
python-dotenv==1.0.1
SQLAlchemy==2.0.28
typing_extensions==4.10.0
